refactor:removed duplicated local assert200HtmlResponse functions and switched to shared import
