+++ ws-nextjs-app/playwright/pageTypes/avEmbed/avEmbed.spec.ts
@@ -1,24 +1,11 @@
-import { test, expect, type APIRequestContext } from '@playwright/test';
+import { test } from '@playwright/test';
 import { avEmbedSuites } from './suites';
 import {
   appEnvFromProcess,
   baseURL,
   shouldRunForEnv,
 } from '../../utilities/env';
-
-const assert200HtmlResponse = async ({
-  request,
-  path,
-}: {
-  request: APIRequestContext;
-  path: string;
-}) => {
-  const response = await request.get(`${baseURL}${path}`);
-  const contentType = response.headers()['content-type'] || '';
-
-  expect(response.status()).toBe(200);
-  expect(contentType).toContain('text/html');
-};
+import assert200HtmlResponse from '../../utilities/response';
 
 test.describe('avEmbed', () => {
   avEmbedSuites.canonical.forEach(testSuite => {
@@ -32,7 +19,11 @@ test.describe('avEmbed', () => {
             `Skipped for APP_ENV=${appEnvFromProcess}`,
           );
 
-          await assert200HtmlResponse({ request, path: testSuite.path });
+          await assert200HtmlResponse({
+            request,
+            path: testSuite.path,
+            baseURL,
+          });
         });
       });
     });
