+++ ws-nextjs-app/playwright/pageTypes/livePage/livePage.spec.ts
@@ -1,4 +1,4 @@
-import { test, expect, type APIRequestContext } from '@playwright/test';
+import { test, expect } from '@playwright/test';
 import { livePageSuites } from './suites';
 import {
   appEnvFromProcess,
@@ -12,20 +12,7 @@ import {
   assertDropdownNavigationComponentView,
   assertDropdownNavigationComponentClick,
 } from '../../specialFeatures/atiAnalytics/assertions';
-
-const assert200HtmlResponse = async ({
-  request,
-  path,
-}: {
-  request: APIRequestContext;
-  path: string;
-}) => {
-  const response = await request.get(`${baseURL}${path}`);
-  const contentType = response.headers()['content-type'] || '';
-
-  expect(response.status()).toBe(200);
-  expect(contentType).toContain('text/html');
-};
+import assert200HtmlResponse from '../../utilities/response';
 
 test.describe('livePage', () => {
   const allSuites = [...livePageSuites.canonical];
@@ -44,7 +31,11 @@ test.describe('livePage', () => {
             `Skipped for APP_ENV=${appEnvFromProcess}`,
           );
 
-          await assert200HtmlResponse({ request, path: testSuite.path });
+          await assert200HtmlResponse({
+            request,
+            path: testSuite.path,
+            baseURL,
+          });
         });
 
         test.describe(`${suiteName} Live Page Tests`, () => {
